Extract current kill and yank to Yanker
